Update README.md
One of the simplest projects to learn about file systems in C++ for first semester students
It show in a simple way the basics behind loginsystem's logic.
This project involves a user registration process by asking username and password.
Upon successful registration, a user file is created with the credentials. If the user does not exist, upon login, an error will be shown.